refactor viewPrimary to use onclick prop #1096
